pyproject.toml
[build-system]										
requires = ["setuptools>=64", "wheel"]                                                  
build-backend = "setuptools.build_meta"                                                 

[project]                                                                               
name = "fibirr-GUI"
version = "1.0.0"                                                                       
description = "GUI to control an Avantes-based fibre-optic spectrometer"
license = "GPL-3.0-or-later"         							
readme = "README.md"                                                                    
requires-python = ">=3.1"
                                                                                        
authors = [                                                                             
  { name = "Jorn Steen", email = "jorn.steen@kemi.uu.se" },
]                                                                                       
                                                                                        
dependencies = [                                                                        
  'numpy>=2.2.2',
  'scipy>=1.15.1',
  'lmfit>=1.3.2',
  'pandas>=2.2.3',
  'PyQt5',
  'matplotlib>=3.10.0',
  'pyserial>=3.5'
]

keywords = [
  "science",
  "spectroscopy"
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Physics",
  "Topic :: Scientific/Engineering :: Chemistry",
  "Topic :: Scientific/Engineering :: Spectroscopy",
  "Topic :: Software Development :: Libraries :: Python Modules"
]

[project.urls]
Homepage = "https://github.com/CrespiLab/fibirr-GUI"

#[project.scripts]
#GUI = "GUImain.main:main" ## need to change GUI.pyw accordingly first

[tool.setuptools]
packages = [
  'avantes',
  'UIs'
]

